# source/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/booking": {
      "post": {
        "tags": ["booking-controller"],
        "operationId": "createBooking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booking/{id}/extend": {
      "post": {
        "summary": "Extend an existing booking",
        "tags": ["booking-controller"],
        "operationId": "extendBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Booking ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["extraNights"],
                "properties": {
                  "extraNights": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking extended successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "400": {
            "description": "Invalid extraNights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "extraNights must be a positive integer"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Booking not found"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: unit is occupied in the extension window",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Extension not possible; unit is taken"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": ["booking-controller"],
        "operationId": "helloWorld",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookingInput": {
        "type": "object",
        "required": ["guestName", "unitID", "checkInDate", "numberOfNights"],
        "properties": {
          "guestName": {
            "type": "string"
          },
          "unitID": {
            "type": "string"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "numberOfNights": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "guestName": {
            "type": "string"
          },
          "unitID": {
            "type": "string"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "numberOfNights": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "id",
          "guestName",
          "unitID",
          "checkInDate",
          "numberOfNights"
        ]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      }
    }
  }
}